requirements.txt
appdirs==1.4.4
asgiref==3.4.0
aspy.refactor-imports==2.2.0
bcrypt==3.2.0
black==21.6b0
cached-property==1.5.2
cffi==1.14.5
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.115.8
gunicorn==20.0.4
greenlet==3.1.1
h11==0.12.0
idna==3.2
importlib-metadata==4.6.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
#psycopg2==2.9.1
psycopg2-binary==2.9.10
pycparser==2.20
pydantic==1.10.21
python-dotenv==0.18.0
regex==2021.7.6
reorder-python-imports==2.5.0
six==1.16.0
SQLAlchemy==1.4.20
starlette==0.45.3
toml==0.10.2
typed-ast==1.5.5
typing-extensions==4.12.2
uvicorn==0.17.6
zipp==3.4.1
google-cloud-pubsub
sendgrid==6.9.0
python-jose
